requirements.txt
## python version: >= 3.10.5
numpy>=1.18
attrs>=23.1.0
typing-extensions>=4.4.0
python-dotenv>=0.21.0
tqdm>=4.64.1
psutil>=5.9.5

-i https://${OWM_GIT_TOKEN}:${OWM_GIT_TOKEN_PASS}@gitlab.openweathermap.org/api/v4/groups/713/-/packages/pypi/simple
deker-local-adapters==0.0.5
